Add abstract sender utility and sync sender now uses it.
This abstract sender is added to allow code reuse regarding
the notifications to listener handling.
Also, this class allows simplified implementations of senders.
