+++ rollbar-java/src/main/java/com/rollbar/notifier/sender/SyncSender.java
@@ -1,10 +1,8 @@
 package com.rollbar.notifier.sender;
 
 import com.rollbar.api.payload.Payload;
-import com.rollbar.notifier.sender.exception.SenderException;
 import com.rollbar.notifier.sender.json.JsonSerializer;
 import com.rollbar.notifier.sender.json.JsonSerializerImpl;
-import com.rollbar.notifier.sender.listener.SenderListener;
 import com.rollbar.notifier.sender.result.Result;
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -15,13 +13,11 @@ import java.net.HttpURLConnection;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
-import java.util.List;
 
 /**
  * Synchronous implementation of the {@link Sender sender}.
  */
-public class SyncSender implements Sender {
+public class SyncSender extends AbstractSender {
 
   public static final String DEFAULT_API_ENDPOINT = "https://api.rollbar.com/api/1/item/";
 
@@ -33,42 +29,16 @@ public class SyncSender implements Sender {
 
   private final String accessToken;
 
-  private final List<SenderListener> listeners;
-
   SyncSender(Builder builder) {
     this.url = builder.url;
     this.jsonSerializer = builder.jsonSerializer;
     this.accessToken = builder.accessToken;
-
-    this.listeners = new ArrayList<>();
   }
 
   @Override
-  public void send(Payload payload) {
-    try {
-      String json = jsonSerializer.toJson(payload);
-      Result result = send(json);
-      notifyResult(payload, result);
-    } catch (Exception e) {
-      notifyError(payload, new SenderException(e));
-    }
-  }
-
-  @Override
-  public void addListener(SenderListener listener) {
-    this.listeners.add(listener);
-  }
-
-  private void notifyResult(Payload payload, Result result) {
-    for (SenderListener listener : listeners) {
-      listener.onResult(payload, result);
-    }
-  }
-
-  private void notifyError(Payload payload, Exception error) {
-    for (SenderListener listener : listeners) {
-      listener.onError(payload, error);
-    }
+  public Result doSend(Payload payload) throws Exception {
+    String json = jsonSerializer.toJson(payload);
+    return send(json);
   }
 
   private Result send(String body) throws IOException {
